refactor: moved validate function to stay inside middleware
